save to and clear localstorage
Use uppercase naming style for mutations as this is one of many conventions, this helps to distinguish mutations from actions.
Call actions from the components instead of mutations. Follow each action by another action that refreshes the localstorage.
The refreshing mechanism can be refactored to not set and remove every item but only the one that is actually changed.
The change event of the input components isn't as verbose as the input event but therefore is triggered only when losing focus.
Cannot yet remove localstorage items by deleting the input value in the form (bug).